createNewChannel() is in CreateNewCh
