+++ frontend/src/components/chat/channelBox/ChannelBox.tsx
@@ -1,12 +1,12 @@
 import styles from './ChannelBox.module.css'
 import CreateNewCh from './CreateNewCh'
 import ChannelList from './ChannelList'
-import { CreateChannel } from '../../../types/CreateChannel'
 import { Channel } from '../../../types/Channel'
+import { Socket } from 'socket.io-client'
 
 interface ChannelBoxProps {
     allChan: Channel[]
-    handleCreation: (channel: CreateChannel) => void
+    socket: Socket | undefined
     deleteChannel: (channelId: number) => void
     leaveChannel: (channelId: number) => void
     joinChannel: (channelId: number, password: string) => void
@@ -16,7 +16,7 @@ interface ChannelBoxProps {
 const ChannelBox = (props: ChannelBoxProps) => {
     return (
         <div className={styles.channelbox}>
-            <CreateNewCh handleCreation={props.handleCreation} />
+            <CreateNewCh socket={props.socket} />
             <ChannelList
                 allChan={props.allChan}
                 deleteChannel={props.deleteChannel}
