+++ frontend/src/components/chat/channelBox/ChannelList.tsx
@@ -58,6 +58,8 @@ const ChannelList = (props: ChannelListProps) => {
         )
     }
 
+  
+
     return (
         <div className={styles.listsContainer}>
             <div className={styles.list}>
